backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
PyJWT==2.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
aiofiles==23.2.1
pillow==10.1.0
pdfplumber==0.11.0
pytesseract==0.3.10
openpyxl==3.1.2
python-docx==0.8.11
httpx==0.27.0
python-dateutil==2.8.2
